Http book fetch now works
