+++ pages/book/BookLists/dummyBookLists.js
@@ -1,0 +1,15 @@
+export const fetchBookList = async () => {
+    return [
+        {
+            title : "Favoritter",
+            id : "1"
+        },
+        {
+            title : "Vil gerne købe",
+            id : "2"
+        },
+        {
+            title : "Interessante",
+            id : "3"
+        }]
+}
