+++ pages/book/bookDetails.css
@@ -170,6 +170,7 @@
     line-height: 17px;
     text-align: justify;
     color: #6A6A6A;
+    margin-top:1rem;
 }
 
 #book-rec-title{
